Move 'setup bash' integration tests to own module
